requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
playwright==1.40.0
requests==2.31.0
ffmpeg-python==0.2.0
aiohttp==3.9.1
aiofiles==23.2.1
